+++ lib/spina/conferences.rb
@@ -4,7 +4,6 @@ require 'spina'
 require 'opal-rails'
 require 'opal-sprockets'
 require 'opal-browser'
-require 'active_record/pg_interval_rails_5_1'
 require 'dotenv-rails'
 require 'uglifier'
